pyproject.toml
[tool.poetry]
name = "hse_slr"
version = "0.1.0"
description = "This Python package is designed for the recognition of Russian sign language and also comes equipped with tools for recognizing various types of body language."
authors = ["asvistunov <aleksandrsvistunov4@gmail.com>", "utnasun <umuniev@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
opencv-python = "4.10.0.82"
aiogram = "^3.3.0"
python-dotenv = "^1.0.0"
sqlalchemy = "^2.0.25"
psycopg2-binary = "^2.9.9"
prettytable = "^3.9.0"
onnx = "1.14.0"
einops = "0.6.1"
pytest = "^8.0.0"
coverage = "^7.4.1"
flake8 = "^7.0.0"
aiogram-tests = {git = "https://github.com/Like6po/aiogram_tests.git"}
pytest-asyncio = "^0.23.5"
ipywidgets = "^8.1.2"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.28.0"}
python-multipart = "^0.0.9"
redis = "^5.0.3"
fastapi-cache2 = "^0.2.1"
pre-commit = "^3.7.0"
timm = "^0.9.16"
pillow = "^10.3.0"
torch = "^2.3.0"
torchvision = "^0.18.0"
loguru = "0.7.0"
omegaconf = "2.3.0"
onnxruntime = "1.14.1"
numpy = "^1.0.0"

[tool.poetry.group.dev.dependencies]
matplotlib = "^3.8.2"
pandas = "^2.1.4"
tqdm = "^4.66.1"
ipykernel = "^6.28.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
